+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Add faces to `helm-projectile-browse-dirty-projects`.
 * Add autoloads to `helm-projectile` commands defined with `helm-projectile-command`.
 * Make `helm-projectile-ag` to use common ignore list
+* Don't add ** to ignored globs in helm-projectile-rg
 
 ## 1.2.0 (2025-07-21)